Rebased code onto event
