+++ Components/Events/views/EventDetailsScreen.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { View, Text, Pressable } from 'react-native';
+import Ionicons from 'react-native-vector-icons/Ionicons';
+import { useNavigation } from '@react-navigation/native';
+import styles from '../styles';
+
+const EventDetailsScreen = () => {
+  const navigation = useNavigation();
+
+  return (
+    <View style={styles.container}>
+      <Pressable
+        style={styles.backButton}
+        onPress={() => navigation.goBack()}
+      >
+        <Ionicons name="arrow-back" size={24} color="#000" />
+      </Pressable>
+      <Text style={styles.detailsTitle}>Event Details</Text>
+    </View>
+  );
+};
+
+export default EventDetailsScreen;
